investigating if the bug is with asm ClassReader
on the issue tracker asm ask if javap works fine on the class
it does, I was able to unzip calicte and run java -p on
SqlFunctions.class

I'm trying to understand if it's an asm bug, or we do something in
jarjar to corrupt the bytecode
